Move presentAssistantMessage into its own module (#3345)
