+++ jest.config.js
@@ -14,7 +14,6 @@ module.exports = {
 					allowJs: true,
 				},
 				diagnostics: false,
-				isolatedModules: true,
 			},
 		],
 	},
