+++ src/activate/index.ts
@@ -2,3 +2,4 @@ export { handleUri } from "./handleUri"
 export { registerCommands } from "./registerCommands"
 export { registerCodeActions } from "./registerCodeActions"
 export { registerTerminalActions } from "./registerTerminalActions"
+export { CodeActionProvider } from "./CodeActionProvider"
