+++ src/core/assistant-message/parseAssistantMessage.ts
